requirements.txt
flask==2.0.1
azure-cosmos==4.3.1
python-dotenv==1.0.0
requests==2.31.0
authlib==1.2.0
Flask-Limiter==3.3.0
tenacity==8.2.2

# flask==2.0.1
# azure-cosmos==4.2.0
# python-dotenv==0.19.0
# requests==2.26.0
# authlib==0.15.4
# jinja2==3.0.3
# werkzeug == 2.2.2